requirements.txt
pillow==11.2.1
requests==2.32.4
tidalapi==0.8.3